% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2h5.r
\name{csv2h5}
\alias{csv2h5}
\title{csv2h5}
\usage{
csv2h5(file, h5out, dataset = NULL, header = "auto",
  format = "column", compression = 4, yolo = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{file}{Input file.}

\item{h5out}{Output file.}

\item{dataset}{Name of the data within the HDF5 file. If none is supplied, then this will be
inferred from the input file name.}

\item{header}{TODO}

\item{format}{One of \code{column} or \code{compound}.}

\item{compression}{HDF5 compression level. An integer between 0 (least) to 9 (most).}

\item{yolo}{Do you want to skip input file checks? Faster, but dangerous.}

\item{verbose}{TODO}

\item{...}{Additional arguments passed to \code{fread()}. Can not include \code{file},
\code{skip}, \code{nrows}, \code{verbose}, \code{showProgress}, or
\code{data.table}.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Convert a csv file to HDF5 dataset.
}
\details{
TODO
}
