% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_h5df.r
\name{is_h5df}
\alias{is_h5df}
\title{Check if an HDF5 dataset is dataframe-like.}
\usage{
is_h5df(h5in, dataset = NULL)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{Name of the data within the HDF5 file. If none is supplied, then this will be
inferred from the input file name.}
}
\value{
Returns \code{TRUE} on success, \code{FALSE} otherwise.
}
\description{
Check if an HDF5 dataset is dataframe-like.
}
\details{
TODO
}
