% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_h5df.r
\name{summarize_h5df}
\alias{summarize_h5df}
\title{summarize_h5df}
\usage{
summarize_h5df(h5in, dataset = NULL, colnames = FALSE)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{Name of the data within the HDF5 file. If none is supplied, then this will be
inferred from the input file name.}

\item{colnames}{TODO}
}
\value{
TODO
}
\description{
TODO
}
