% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5df.r
\name{write_h5df}
\alias{write_h5df}
\title{write_h5df}
\usage{
write_h5df(x, file, dataset = NULL, format = "column",
  compression = 4)
}
\arguments{
\item{x}{Input dataset.}

\item{file}{Output file.}

\item{dataset}{Name of the data within the HDF5 file. If none is supplied, then this will be
inferred from the input file name.}

\item{format}{One of \code{column} or \code{compound}.}

\item{compression}{HDF5 compression level. An integer between 0 (least) to 9 (most).}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
write_h5df
}
\details{
The input dataset should be dataframe-like (one of \code{data.frame},
\code{data.table}, or \code{tbl_df}), or coercible to a \code{data.frame} via
\code{as.data.frame()}.
}
