% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2h5.r
\name{csv2h5}
\alias{csv2h5}
\title{csv2h5}
\usage{
csv2h5(csvfile, csvdir = NULL, h5out, dataset, format = "column",
  compression = 0, stringsAsFactors = FALSE, yolo = FALSE)
}
\arguments{
\item{csvfile}{Input file.}

\item{csvdir}{TODO}

\item{dataset}{TODO}

\item{format}{TODO}

\item{compression}{HDF5 compression level. An integer, 0 (least compression) to 9 (most
compression).}

\item{stringsAsFactors}{TODO}

\item{yolo}{TODO}

\item{h5file}{Output file.}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Convert a csv to HDF5 dataset.
}
\details{
TODO
}

