% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2h5.r
\name{csv2h5}
\alias{csv2h5}
\alias{csv2h5}
\alias{dir2h5}
\title{csv2h5}
\usage{
csv2h5(file, h5out, dataset = NULL, format = "column", compression = 0,
  stringsAsFactors = FALSE, yolo = FALSE, verbose = FALSE)

dir2h5(csvdir, h5out, dataset = NULL, format = "column", compression = 0,
  stringsAsFactors = FALSE, yolo = FALSE, verbose = FALSE)
}
\arguments{
\item{file}{Input file.}

\item{h5out}{Output file.}

\item{dataset}{TODO}

\item{format}{TODO}

\item{compression}{HDF5 compression level. An integer, 0 (least compression) to 9 (most
compression).}

\item{stringsAsFactors}{TODO}

\item{yolo}{TODO}

\item{verbose}{TODO}

\item{csvdir}{TODO}
}
\value{
Invisibly returns \code{TRUE} on success.
}
\description{
Convert a csv file or a directory of csv files to HDF5 dataset.
}
\details{
TODO
}
