% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_h5df.r
\name{is_h5df}
\alias{is_h5df}
\title{is_h5df}
\usage{
is_h5df(h5in, dataset = NULL)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{TODO}
}
\value{
Returns \code{TRUE} on success, \code{FALSE} otherwise.
}
\description{
Check if an HDF5 dataset is dataframe-like.
}
\details{
TODO
}
\examples{
library(hdfio)

f = system.file("exampledata/pytables_table.h5", package="hdfio")
is_h5df(f, "mydata")
is_h5df(f, "/")

f = system.file("exampledata/not_h5df.h5", package="hdfio")
is_h5df(f, "mydata")

}
\seealso{
\code{\link{is_h5df}}
}

