% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5df.r
\name{read_h5df}
\alias{read_h5df}
\title{read_h5df}
\usage{
read_h5df(h5in, dataset = NULL, rows = NULL, cols = NULL,
  strings = TRUE, verbose = FALSE)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{Name of the data within the HDF5 file. If none is supplied, then this will be
inferred from the input file name.}

\item{rows}{TODO}

\item{cols}{TODO}

\item{strings}{Only available for 'hdfio_columns' format files. Should string columns be read?}

\item{verbose}{TODO}
}
\value{
A dataframe.
}
\description{
TODO
}
\details{
TODO
}
