% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_h5.r
\name{read_h5df}
\alias{read_h5df}
\title{read_h5df}
\usage{
read_h5df(h5in, dataset)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{Dataset in input file to read.}
}
\value{
A dataframe.
}
\description{
TODO
}
\examples{
library(hdfio)

h5in = system.file("exampledata/from_pandas.h5", package="hdfio")
df = read_h5df(h5in, "mydata")
df

}
\seealso{
\code{\link{is_h5df}}
}
