% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_h5df.r
\name{summarize_h5df}
\alias{summarize_h5df}
\title{summarize_h5df}
\usage{
summarize_h5df(h5in, dataset = NULL, colnames = FALSE)
}
\arguments{
\item{h5in}{Input file.}

\item{dataset}{TODO}

\item{colnames}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
library(hdfio)
h5in = system.file("exampledata/pytables_table.h5", package="hdfio")
summarize_h5df(h5in)

}
