% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5df.r
\name{write_h5df}
\alias{write_h5df}
\title{write_h5df}
\usage{
write_h5df(x, file, dataset = NULL, format = "column", compression = 4)
}
\arguments{
\item{x}{Input dataset (dataframe or datatable)}

\item{file}{Output file.}

\item{dataset}{Dataset in input file to read or \code{NULL}. In the latter case (e.g. \code{NULL}), the dataset will be contained 
within a group named as the input dataset.}

\item{format}{Method chosen for writing out h5 file.  If \code{column}, each column of the input dataset is written 
out on disk as x_i with "i" being an arbitrary column index, ranging as intengers from 1:ncol(dataframe). 
If \code{compound}, the entire input dataset is written out on disk
as a complete dataframe}

\item{compression}{HDF5 compression level. An integer, 0 (least compression) to 9 (most compression).  Default is
\code{compression} = 4.}
}
\value{
A dataframe.
}
\description{
write_h5df
}
\details{
TODO
}
\examples{
#Example 1
library(hdfio)
df = data.frame(x=seq(1:5), y = c(runif(5)), z= c("Peter", "Amber", "John", "Lindsey", "Steven"))
(df)

#Writing out data in column format to a hdf5 group "data", where each variable is indexed as x1,x2, and x3 

write_h5df(x = df, file = paste(tempdir(), "example.h5", sep="/"), dataset = "data", format = "column", compression=4) 
#To verify, we can read the data back in.
#Result
read_h5df(paste(tempdir(), "example.h5", sep="/"), "data")


#Example 2
#Write dataframe (df) out in compound format 
write_h5df(x = df, file = paste(tempdir(), "example2.h5", sep="/"), dataset = "data", format = "compound", compression=4) 

#To verify, we read the data back in.
#Result
read_h5df(paste(tempdir(), "example2.h5", sep="/"))



}
\seealso{
\code{\link{read_h5df}}
}
