% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5df.r
\name{write_h5df}
\alias{write_h5df}
\title{write_h5df}
\usage{
write_h5df(x, file, dataset, stringsAsFactors = TRUE, format = "column")
}
\arguments{
\item{x}{Input dataset (dataframe or datatable)}

\item{file}{Output file.}

\item{dataset}{Dataset in input file to read or \code{NULL}. In the latter case, TODO}

\item{stringsAsFactors}{TODO}

\item{format}{TODO}
}
\value{
A dataframe.
}
\description{
TODO
}
\examples{
library(hdfio)

# TODO

}
\seealso{
\code{\link{read_h5df}}
}

